requirements.txt
langchain
langchain-classic
langchain-community
langchain-text-splitters
streamlit
langchain-google-genai
langsmith
langchain-chroma
chromadb
fastapi
uvicorn
pypdf
python-dotenv